Separated ascii shader into distinct brightness and edge ascii shaders
